# UnrealPluginManager.Server/Source/UnrealPluginManager.Server/openapi-spec.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Unreal Plugin Manager API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
    "/plugins": {
      "get": {
        "tags": [
          "Plugins"
        ],
        "summary": "Retrieves a paginated list of plugin overviews based on the specified filter and pagination settings.",
        "operationId": "getPlugins",
        "parameters": [
          {
            "name": "match",
            "in": "query",
            "description": "A wildcard string used to filtered plugins by name. Defaults to \"*\".",
            "schema": {
              "type": "string",
              "default": "*"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The page number to retrieve.",
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The number of items to retrieve per page.",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginOverviewPage"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Plugins"
        ],
        "summary": "Submits a plugin for processing by uploading source code and a collection of binaries.",
        "operationId": "submitPlugin",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "submission": {
                    "type": "string",
                    "description": "An object containing the plugin's source code file and associated binaries for submission.",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "submission": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginVersionDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "CanSubmitPlugin"
            ]
          },
          {
            "apiKey": [
              "CanSubmitPlugin"
            ]
          }
        ]
      }
    },
    "/plugins/latest": {
      "get": {
        "tags": [
          "Plugins"
        ],
        "summary": "Retrieves a paginated list of the latest plugin versions filtered by the specified criteria.",
        "operationId": "getLatestVersions",
        "parameters": [
          {
            "name": "match",
            "in": "query",
            "description": "A wildcard string used to filter plugins by name. Defaults to \"*\".",
            "schema": {
              "type": "string",
              "default": "*"
            }
          },
          {
            "name": "versionRange",
            "in": "query",
            "description": "The semantic version range to filter the plugin versions. Defaults to all release versions.",
            "schema": {
              "type": "string",
              "example": ">=1.0.0"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The page number to retrieve.",
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The number of items to retrieve per page.",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginVersionInfoPage"
                }
              }
            }
          }
        }
      }
    },
    "/plugins/{engineVersion}/submit": {
      "post": {
        "tags": [
          "Plugins"
        ],
        "summary": "Adds a plugin by uploading a plugin file and specifying the target Unreal Engine version.",
        "operationId": "addPlugin",
        "parameters": [
          {
            "name": "engineVersion",
            "in": "path",
            "description": "The target Unreal Engine version for which the plugin is being added.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "pluginFile": {
                    "type": "string",
                    "description": "The uploaded plugin file in a valid format.",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "pluginFile": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginVersionDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "CanSubmitPlugin"
            ]
          },
          {
            "apiKey": [
              "CanSubmitPlugin"
            ]
          }
        ]
      }
    },
    "/plugins/{pluginId}/latest": {
      "get": {
        "tags": [
          "Plugins"
        ],
        "summary": "Retrieves detailed information about the latest version of the specified plugin,\r\noptionally constrained by a version range.",
        "operationId": "getLatestVersion",
        "parameters": [
          {
            "name": "pluginId",
            "in": "path",
            "description": "The unique identifier of the plugin to retrieve the latest version for.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "version",
            "in": "query",
            "description": "An optional version range to filter the plugin's versions. Defaults to all released versions.",
            "schema": {
              "type": "string",
              "example": ">=1.0.0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginVersionInfo"
                }
              }
            }
          }
        }
      }
    },
    "/plugins/{pluginId}/latest/dependencies": {
      "get": {
        "tags": [
          "Plugins"
        ],
        "summary": "Retrieves the dependency tree for a specified plugin.",
        "operationId": "getDependencyTree",
        "parameters": [
          {
            "name": "pluginId",
            "in": "path",
            "description": "The unique identifier of the plugin whose dependency tree is to be retrieved.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "An optional version range used to filter dependencies for the plugin.",
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": ">=1.0.0"
              }
            },
            "text/json": {
              "schema": {
                "type": "string",
                "example": ">=1.0.0"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string",
                "example": ">=1.0.0"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PluginSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/plugins/{pluginId}/latest/{engineVersion}/download": {
      "get": {
        "tags": [
          "Plugins"
        ],
        "summary": "Downloads a plugin file as a ZIP archive for the specified plugin, engine version, and target platforms.",
        "operationId": "downloadLatestPlugin",
        "parameters": [
          {
            "name": "pluginId",
            "in": "path",
            "description": "The unique identifier of the plugin to be downloaded.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "engineVersion",
            "in": "path",
            "description": "The Unreal Engine version for which the plugin file is requested.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "targetVersion",
            "in": "query",
            "description": "The semantic version range that specifies the version of the plugin to target. Defaults to all release versions if not specified.",
            "schema": {
              "type": "string",
              "example": ">=1.0.0"
            }
          },
          {
            "name": "platforms",
            "in": "query",
            "description": "The collection of target platforms for which the plugin file is compatible.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "separated",
            "in": "query",
            "description": "A boolean value indicating whether to separate the plugin files by platform. Defaults to false.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
    "/plugins/{pluginId}/{versionId}/download/{engineVersion}": {
      "get": {
        "tags": [
          "Plugins"
        ],
        "summary": "Downloads the specified version of a plugin as a ZIP file for the specified Unreal Engine version and target platforms.",
        "operationId": "downloadPluginVersion",
        "parameters": [
          {
            "name": "pluginId",
            "in": "path",
            "description": "The unique identifier of the plugin to download.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "versionId",
            "in": "path",
            "description": "The unique identifier of the plugin version to download.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "engineVersion",
            "in": "path",
            "description": "The version of Unreal Engine compatible with the plugin.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "platforms",
            "in": "query",
            "description": "The collection of target platforms for the plugin.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "separated",
            "in": "query",
            "description": "A boolean value indicating whether to separate the plugin files by platform. Defaults to false.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/plugins/{pluginId}/{versionId}/download/source": {
      "get": {
        "tags": [
          "Plugins"
        ],
        "summary": "Downloads the source code of a specific plugin version as a zip file.",
        "operationId": "downloadPluginSource",
        "parameters": [
          {
            "name": "pluginId",
            "in": "path",
            "description": "The unique identifier of the plugin to download.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "versionId",
            "in": "path",
            "description": "The unique identifier of the specific version of the plugin to download.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/plugins/{pluginId}/{versionId}/download/{engineVersion}/{platform}/binaries": {
      "get": {
        "tags": [
          "Plugins"
        ],
        "summary": "Downloads the binary files of a specified plugin for a given version, engine version, and platform.",
        "operationId": "downloadPluginBinaries",
        "parameters": [
          {
            "name": "pluginId",
            "in": "path",
            "description": "The unique identifier of the plugin whose binaries are being downloaded.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "versionId",
            "in": "path",
            "description": "The unique identifier of the plugin version to download binaries for.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "engineVersion",
            "in": "path",
            "description": "The engine version for which the plugin binaries are compatible.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "platform",
            "in": "path",
            "description": "The platform for which the plugin binaries are compiled.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PluginOverviewPage": {
        "required": [
          "pageNumber",
          "totalPages",
          "pageSize",
          "count",
          "items"
        ],
        "type": "object",
        "properties": {
          "pageNumber": {
            "minimum": 1,
            "type": "integer",
            "description": "The current page number of the result set.",
            "format": "int32"
          },
          "totalPages": {
            "minimum": 1,
            "type": "integer",
            "description": "The total number of pages in the result set.",
            "format": "int32"
          },
          "pageSize": {
            "minimum": 1,
            "type": "integer",
            "description": "The number of items on each page.",
            "format": "int32",
            "example": 10
          },
          "count": {
            "minimum": 0,
            "type": "integer",
            "description": "The total number of items available in the result set.",
            "format": "int32",
            "example": 1
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PluginOverview"
            },
            "description": "An array of items on the current page."
          }
        },
        "description": "Represents the output of a paged query of PluginOverview."
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/kc/realms/unreal-plugin-manager/protocol/openid-connect/auth",
            "tokenUrl": "/kc/realms/unreal-plugin-manager/protocol/openid-connect/token",
            "scopes": {
              "CanSubmitPlugin": "User is a contributor on the given plugin and has submit privileges",
              "CanEditPlugin": "User is a contributor on the given plugin and has edit privileges"
            }
          }
        }
      },
      "apiKey": {
        "type": "apiKey",
        "name": "X-API-Key",
        "in": "header"
      }
    }
  }
}